nixos/test-driver: refactor in preparation of tmux test driver
