+++ nixos/lib/test-driver/lib/common.py
@@ -1,0 +1,6 @@
+from typing import Any
+import sys
+
+
+def eprint(*args: object, **kwargs: Any) -> None:
+    print(*args, file=sys.stderr, **kwargs)
